chore(lint): adopt v2 lint-first selectors (N/SIM/RUF/ERA/T20) + pyright strict on src/ (#2)
* chore: gitignore .claude/ to prevent agent-memory leakage

Per-agent persistent memory is canonical at
/workspaces/ocr-container/.claude/agent-memory/<agent>/, never inside
this repo's tree. Historically agents would land memory here when
their cwd was wrong; this gitignore stops the leakage from getting
committed and matches a workspace-wide policy update.


* docs(dev-local): spec dev-local-aware upgrade-deps flow + roadmap (t1)

Captures the workspace-wide hazard where `make upgrade-deps` ends with
`uv sync --group all-dev` and silently reverts a dev-local venv
(editable pd-book-tools, editable sibling doctr, operator-installed
CUDA torch wheels) to canonical/CPU. Trainer is the most GPU-sensitive
repo in the workspace, so the GPU-extras restore step is called out
explicitly as a thing dev-local mode must preserve.

Adds docs/specs/dev-local-upgrade-flow.md (detection cascade, refuse-
with-message default, sibling upgrade-deps-local recipe) and a (t1)
tooling milestone in docs/ROADMAP.md. Also adopts the workspace-
standard .markdownlint-cli2.jsonc (matches pd-book-tools et al.) so
the previously-untracked docs/*.md content lints clean on first commit.

No Makefile changes in this pass — spec only.


* docs(roadmap): spec local-mode port auto-select + stable bookmarks + in-UI URL

Add (t2) tooling milestone and companion spec for the NiceGUI training
UI: try persisted-last-port → default → port=0; honor explicit --port
with loud failure; persist last bound port for stable bookmarks; surface
the URL inside the running UI for operators who closed the console.

Cross-links sibling items in pd-prep-for-pgdp (b23b913),
pd-ocr-labeler-spa (b956275), and pd-ocr-labeler (parallel).


* docs: add CLAUDE.md with compact AI policy


* docs: add datasets spec, layout training spec, glyph classifier specs, code review

- DATASETS.md: Hugging Face Datasets integration architecture spec
- SPEC-layout-training.md: layout model training specification
- TOP-50-LABELING-TARGETS.md: labeling priority list
- docs/review/code-review.md: deep code review findings (train_recog/detect bugs)
- docs/specs/glyph-annotation-eval-slicing.md: glyph eval slicing spec
- docs/specs/glyph-feature-classifier.md: glyph feature classifier spec
- .gitignore: exclude ml-training/ ml-validation/ dirs
- README.md: add roadmap section linking ROADMAP.md and DATASETS.md
- .claudeignore: exclude build/training dirs from Claude context
- .markdownlint-cli2.jsonc: allow sibling-level duplicate headings


* chore(make): guard pre-commit install with [ -f .git/hooks/pre-commit ] [pilot-feedback]

Matches pd-book-tools 57eeeda; fixes EPERM when claude-bot runs `make`
after vscode already installed the hook.

* chore(lint): adopt v2 lint-first selectors (N/SIM/RUF/ERA/T20) + pyright strict on src/

Adds N, SIM, RUF, ERA, T20 to [tool.ruff.lint] select and adds a
[tool.pyright] section with basic mode globally and strict mode for src/.

All 14 new findings fixed in place:
- N817/N812: noqa with justification (PyTorch/doctr conventional aliases)
- SIM105: replace try/except/pass with contextlib.suppress (3 sites)
- SIM211: simplify `False if cond else True` -> `not cond` (2 sites)
- T201: replace print() with logging.info/logger.debug (4 sites)
- RUF046: remove redundant int() around math.ceil() (1 site)


---------

+++ src/pd_ocr_trainer/utils.py
@@ -8,7 +8,7 @@ def plot_samples(images, targets):
     # Unnormalize image
     num_samples = min(len(images), 12)
     num_cols = min(len(images), 4)
-    num_rows = int(math.ceil(num_samples / num_cols))
+    num_rows = math.ceil(num_samples / num_cols)
     _, axes = plt.subplots(num_rows, num_cols, figsize=(20, 5))
     for idx in range(num_samples):
         img = (255 * images[idx].numpy()).round().clip(0, 255).astype(np.uint8)
